feat: add app new pipeline
