+++ shell/app/common/utils/go-to.tsx
@@ -196,7 +196,9 @@ export enum pages {
   repo = '/{orgName}/dop/projects/{projectId}/apps/{appId}/repo',
   appMr = '/{orgName}/dop/projects/{projectId}/apps/{appId}/repo/mr/open/{mrId}',
   appOpenMr = '/{orgName}/dop/projects/{projectId}/apps/{appId}/repo/mr/open',
-  pipelineRoot = '/{orgName}/dop/projects/{projectId}/apps/{appId}/pipeline',
+  pipelineRoot = '/{orgName}/dop/projects/{projectId}/apps/{appId}/pipeline/list',
+  appPipelineRecords = '/{orgName}/dop/projects/{projectId}/apps/{appId}/pipeline/records',
+  appPipelineDetail = '/{orgName}/dop/projects/{projectId}/apps/{appId}/pipeline-detail/{pipelineID}',
   appApiDesign = '/{orgName}/dop/projects/{projectId}/apps/{appId}/apiDesign',
   repoBackup = '/{orgName}/dop/projects/{projectId}/apps/{appId}/repo/backup',
   commit = '/{orgName}/dop/projects/{projectId}/apps/{appId}/repo/commit/{commitId}',
